+++ src/redteam/plugins/harmful/utils.ts
@@ -1,0 +1,11 @@
+import type { TestCase } from '../../..';
+
+export function sortTestCasesByCategory(a: TestCase, b: TestCase): number {
+  const categoryComparison = (a?.metadata?.harmCategory || '').localeCompare(
+    b?.metadata?.harmCategory || '',
+  );
+  if (categoryComparison !== 0) {
+    return categoryComparison;
+  }
+  return JSON.stringify(a?.vars || {}).localeCompare(JSON.stringify(b?.vars || {}));
+}
